Update to 1.8.4
We're building our own RPMs in our 3rdparty repo.
Update DDL location to 1.8.4

The reason for the shellout is that the web DDL uses \set variables (for psql
batch) that aren't compatible with a normal db query. The reason for putting it
in a shellout instead of a bash_script resource is that I trust num_tuples more
than cobbled-up shell scripts using psql.